Fix softmax dtype, checkpoint resume path, and pretokenize directory
- MultiHeadAttention softmax: cast to float32 then back to v.dtype (matches TemporalAttentionTok)
- Save "latest" checkpoint to base dir (checkpoints/) for sbatch resume discovery
- Add resume directory reuse to tokenizer script (dynamics already had it)
- Lower step_save_interval default 20K→15K to fit within 1hr preemption grace
- Fix pretokenize_frames.py: look directly in video_dir, not video_dir/frames/

+++ src/ahriuwu/models/transformer_tokenizer.py
@@ -265,7 +265,7 @@ class MultiHeadAttention(nn.Module):
                 mask = mask.unsqueeze(1)  # (B, 1, N, N)
             attn = attn.masked_fill(~mask, float('-inf'))
 
-        attn = F.softmax(attn, dim=-1)
+        attn = F.softmax(attn, dim=-1, dtype=torch.float32).to(v.dtype)
         attn = self.dropout(attn)
 
         out = torch.matmul(attn, v)
